refactor: separate structural plasticity from archiving node to its own class
