+++ nestkernel/archiving_node.cpp
@@ -20,13 +20,6 @@
  *
  */
 
-/**
- * \file archiving_node.cpp
- * Implementation of archiving_node to record and manage spike history
- * \author Moritz Helias, Abigail Morrison
- * \date april 2006
- */
-
 #include "archiving_node.h"
 
 // Includes from nestkernel:
@@ -50,16 +43,11 @@ nest::Archiving_Node::Archiving_Node()
   , tau_minus_triplet_inv_( 1. / tau_minus_triplet_ )
   , max_delay_( 0 )
   , last_spike_( -1.0 )
-  , Ca_t_( 0.0 )
-  , Ca_minus_( 0.0 )
-  , tau_Ca_( 10000.0 )
-  , beta_Ca_( 0.001 )
-  , synaptic_elements_map_()
 {
 }
 
 nest::Archiving_Node::Archiving_Node( const Archiving_Node& n )
-  : Node( n )
+  : Structural_Plasticity_Node( n )
   , n_incoming_( n.n_incoming_ )
   , Kminus_( n.Kminus_ )
   , Kminus_triplet_( n.Kminus_triplet_ )
@@ -70,11 +58,6 @@ nest::Archiving_Node::Archiving_Node( const Archiving_Node& n )
   , max_delay_( n.max_delay_ )
   , trace_( n.trace_ )
   , last_spike_( n.last_spike_ )
-  , Ca_t_( n.Ca_t_ )
-  , Ca_minus_( n.Ca_minus_ )
-  , tau_Ca_( n.tau_Ca_ )
-  , beta_Ca_( n.beta_Ca_ )
-  , synaptic_elements_map_( n.synaptic_elements_map_ )
 {
 }
 
@@ -196,9 +179,9 @@ nest::Archiving_Node::get_history( double t1,
 void
 nest::Archiving_Node::set_spiketime( Time const& t_sp, double offset )
 {
+  Structural_Plasticity_Node::set_spiketime( t_sp, offset );
+
   const double t_sp_ms = t_sp.get_ms() - offset;
-  update_synaptic_elements( t_sp_ms );
-  Ca_minus_ += beta_Ca_;
 
   if ( n_incoming_ )
   {
@@ -236,30 +219,16 @@ nest::Archiving_Node::set_spiketime( Time const& t_sp, double offset )
 void
 nest::Archiving_Node::get_status( DictionaryDatum& d ) const
 {
-  DictionaryDatum synaptic_elements_d;
-  DictionaryDatum synaptic_element_d;
-
   def< double >( d, names::t_spike, get_spiketime_ms() );
   def< double >( d, names::tau_minus, tau_minus_ );
-  def< double >( d, names::Ca, Ca_minus_ );
-  def< double >( d, names::tau_Ca, tau_Ca_ );
-  def< double >( d, names::beta_Ca, beta_Ca_ );
   def< double >( d, names::tau_minus_triplet, tau_minus_triplet_ );
   def< double >( d, names::post_trace, trace_ );
 #ifdef DEBUG_ARCHIVER
   def< int >( d, names::archiver_length, history_.size() );
 #endif
 
-  synaptic_elements_d = DictionaryDatum( new Dictionary );
-  def< DictionaryDatum >( d, names::synaptic_elements, synaptic_elements_d );
-  for ( std::map< Name, SynapticElement >::const_iterator it = synaptic_elements_map_.begin();
-        it != synaptic_elements_map_.end();
-        ++it )
-  {
-    synaptic_element_d = DictionaryDatum( new Dictionary );
-    def< DictionaryDatum >( synaptic_elements_d, it->first, synaptic_element_d );
-    it->second.get( synaptic_element_d );
-  }
+  // add status dict items from the parent class
+  Structural_Plasticity_Node::get_status( d );
 }
 
 void
@@ -268,36 +237,21 @@ nest::Archiving_Node::set_status( const DictionaryDatum& d )
   // We need to preserve values in case invalid values are set
   double new_tau_minus = tau_minus_;
   double new_tau_minus_triplet = tau_minus_triplet_;
-  double new_tau_Ca = tau_Ca_;
-  double new_beta_Ca = beta_Ca_;
   updateValue< double >( d, names::tau_minus, new_tau_minus );
   updateValue< double >( d, names::tau_minus_triplet, new_tau_minus_triplet );
-  updateValue< double >( d, names::tau_Ca, new_tau_Ca );
-  updateValue< double >( d, names::beta_Ca, new_beta_Ca );
 
   if ( new_tau_minus <= 0.0 or new_tau_minus_triplet <= 0.0 )
   {
     throw BadProperty( "All time constants must be strictly positive." );
   }
 
+  Structural_Plasticity_Node::set_status( d );
+
+  // do the actual update
   tau_minus_ = new_tau_minus;
   tau_minus_triplet_ = new_tau_minus_triplet;
   tau_minus_inv_ = 1. / tau_minus_;
   tau_minus_triplet_inv_ = 1. / tau_minus_triplet_;
-
-  if ( new_tau_Ca <= 0.0 )
-  {
-    throw BadProperty( "All time constants must be strictly positive." );
-  }
-  tau_Ca_ = new_tau_Ca;
-
-  if ( new_beta_Ca <= 0.0 )
-  {
-    throw BadProperty(
-      "For Ca to function as an integrator of the electrical activity, beta_ca "
-      "needs to be greater than 0." );
-  }
-  beta_Ca_ = new_beta_Ca;
 
   // check, if to clear spike history and K_minus
   bool clear = false;
@@ -305,38 +259,6 @@ nest::Archiving_Node::set_status( const DictionaryDatum& d )
   if ( clear )
   {
     clear_history();
-  }
-
-  if ( d->known( names::synaptic_elements_param ) )
-  {
-    const DictionaryDatum synaptic_elements_dict = getValue< DictionaryDatum >( d, names::synaptic_elements_param );
-
-    for ( std::map< Name, SynapticElement >::iterator it = synaptic_elements_map_.begin();
-          it != synaptic_elements_map_.end();
-          ++it )
-    {
-      if ( synaptic_elements_dict->known( it->first ) )
-      {
-        const DictionaryDatum synaptic_elements_a = getValue< DictionaryDatum >( synaptic_elements_dict, it->first );
-        it->second.set( synaptic_elements_a );
-      }
-    }
-  }
-  if ( not d->known( names::synaptic_elements ) )
-  {
-    return;
-  }
-  // we replace the existing synaptic_elements_map_ by the new one
-  DictionaryDatum synaptic_elements_d;
-  std::pair< std::map< Name, SynapticElement >::iterator, bool > insert_result;
-
-  synaptic_elements_map_ = std::map< Name, SynapticElement >();
-  synaptic_elements_d = getValue< DictionaryDatum >( d, names::synaptic_elements );
-
-  for ( Dictionary::const_iterator i = synaptic_elements_d->begin(); i != synaptic_elements_d->end(); ++i )
-  {
-    insert_result = synaptic_elements_map_.insert( std::pair< Name, SynapticElement >( i->first, SynapticElement() ) );
-    ( insert_result.first->second ).set( getValue< DictionaryDatum >( synaptic_elements_d, i->first ) );
   }
 }
 
@@ -347,122 +269,7 @@ nest::Archiving_Node::clear_history()
   Kminus_ = 0.0;
   Kminus_triplet_ = 0.0;
   history_.clear();
-  Ca_minus_ = 0.0;
-  Ca_t_ = 0.0;
 }
 
-
-/* ----------------------------------------------------------------
-* Get the number of synaptic_elements
-* ---------------------------------------------------------------- */
-double
-nest::Archiving_Node::get_synaptic_elements( Name n ) const
-{
-  std::map< Name, SynapticElement >::const_iterator se_it;
-  se_it = synaptic_elements_map_.find( n );
-  double z_value;
-
-  if ( se_it != synaptic_elements_map_.end() )
-  {
-    z_value = ( se_it->second ).get_z();
-    if ( ( se_it->second ).continuous() )
-    {
-      return z_value;
-    }
-    else
-    {
-      return std::floor( z_value );
-    }
-  }
-  else
-  {
-    return 0.0;
-  }
-}
-
-int
-nest::Archiving_Node::get_synaptic_elements_vacant( Name n ) const
-{
-  std::map< Name, SynapticElement >::const_iterator se_it;
-  se_it = synaptic_elements_map_.find( n );
-
-  if ( se_it != synaptic_elements_map_.end() )
-  {
-    return se_it->second.get_z_vacant();
-  }
-  else
-  {
-    return 0;
-  }
-}
-
-int
-nest::Archiving_Node::get_synaptic_elements_connected( Name n ) const
-{
-  std::map< Name, SynapticElement >::const_iterator se_it;
-  se_it = synaptic_elements_map_.find( n );
-
-  if ( se_it != synaptic_elements_map_.end() )
-  {
-    return se_it->second.get_z_connected();
-  }
-  else
-  {
-    return 0;
-  }
-}
-
-std::map< Name, double >
-nest::Archiving_Node::get_synaptic_elements() const
-{
-  std::map< Name, double > n_map;
-
-  for ( std::map< Name, SynapticElement >::const_iterator it = synaptic_elements_map_.begin();
-        it != synaptic_elements_map_.end();
-        ++it )
-  {
-    n_map.insert( std::pair< Name, double >( it->first, get_synaptic_elements( it->first ) ) );
-  }
-  return n_map;
-}
-
-void
-nest::Archiving_Node::update_synaptic_elements( double t )
-{
-  assert( t >= Ca_t_ );
-
-  for ( std::map< Name, SynapticElement >::iterator it = synaptic_elements_map_.begin();
-        it != synaptic_elements_map_.end();
-        ++it )
-  {
-    it->second.update( t, Ca_t_, Ca_minus_, tau_Ca_ );
-  }
-  // Update calcium concentration
-  Ca_minus_ = Ca_minus_ * std::exp( ( Ca_t_ - t ) / tau_Ca_ );
-  Ca_t_ = t;
-}
-
-void
-nest::Archiving_Node::decay_synaptic_elements_vacant()
-{
-  for ( std::map< Name, SynapticElement >::iterator it = synaptic_elements_map_.begin();
-        it != synaptic_elements_map_.end();
-        ++it )
-  {
-    it->second.decay_z_vacant();
-  }
-}
-
-void
-nest::Archiving_Node::connect_synaptic_element( Name name, int n )
-{
-  std::map< Name, SynapticElement >::iterator se_it;
-  se_it = synaptic_elements_map_.find( name );
-
-  if ( se_it != synaptic_elements_map_.end() )
-  {
-    se_it->second.connect( n );
-  }
-}
 
 } // of namespace nest
