+++ nestkernel/archiving_node.h
@@ -20,14 +20,6 @@
  *
  */
 
-/**
- * \file archiving_node.h
- * Definition of Archiving_Node which is capable of
- * recording and managing a spike history.
- * \author Moritz Helias, Abigail Morrison
- * \date april 2006
- */
-
 #ifndef ARCHIVING_NODE_H
 #define ARCHIVING_NODE_H
 
@@ -40,6 +32,7 @@
 #include "nest_time.h"
 #include "nest_types.h"
 #include "node.h"
+#include "structural_plasticity_node.h"
 #include "synaptic_element.h"
 
 // Includes from sli:
@@ -51,14 +44,11 @@ namespace nest
 {
 
 /**
- * \class Archiving_Node
- * a node which archives spike history for the purposes of
- * timing dependent plasticity
+ * A node which archives spike history for the purposes of spike-timing
+ * dependent plasticity (STDP)
  */
-class Archiving_Node : public Node
+class Archiving_Node : public Structural_Plasticity_Node
 {
-  using Node::get_synaptic_elements;
-
 public:
   /**
    * \fn Archiving_Node()
@@ -71,61 +61,6 @@ public:
    * Copy Constructor.
    */
   Archiving_Node( const Archiving_Node& );
-  /**
-
-   * \fn double get_Ca_minus()
-   * return the current value of Ca_minus
-   */
-  double get_Ca_minus() const;
-
-  /**
-   * \fn double get_synaptic_elements(Name n)
-   * get the number of synaptic element for the current Node
-   * the number of synaptic elements is a double value but the number of
-   * actual vacant and connected elements is an integer truncated from this
-   * value
-   */
-  double get_synaptic_elements( Name n ) const;
-
-  /**
-   * \fn int get_synaptic_elements_vacant(Name n)
-   * Get the number of synaptic elements of type n which are available
-   * for new synapse creation
-   */
-  int get_synaptic_elements_vacant( Name n ) const;
-
-  /**
-   * \fn int get_synaptic_elements_connected(Name n)
-   * get the number of synaptic element of type n which are currently
-   * connected
-   */
-  int get_synaptic_elements_connected( Name n ) const;
-
-  /**
-   * \fn std::map<Name, double> get_synaptic_elements()
-   * get the number of all synaptic elements for the current Node
-   */
-  std::map< Name, double > get_synaptic_elements() const;
-
-  /**
-   * \fn void update_synaptic_elements()
-   * Change the number of synaptic elements in the node depending on the
-   * dynamics described by the corresponding growth curve
-   */
-  void update_synaptic_elements( double t );
-
-  /**
-   * \fn void decay_synaptic_elements_vacant()
-   * Delete a certain portion of the vacant synaptic elements which are not
-   * in use
-   */
-  void decay_synaptic_elements_vacant();
-
-  /**
-   * \fn void connect_synaptic_element()
-   * Change the number of connected synaptic elements by n
-   */
-  void connect_synaptic_element( Name name, int n );
 
   /**
    * \fn double get_K_value(long t)
@@ -194,12 +129,6 @@ public:
   void get_status( DictionaryDatum& d ) const;
   void set_status( const DictionaryDatum& d );
 
-  /**
-   * retrieve the current value of tau_Ca which defines the exponential decay
-   * constant of the intracellular calcium concentration
-   */
-  double get_tau_Ca() const;
-
 protected:
   /**
    * \fn void set_spiketime(Time const & t_sp, double offset)
@@ -245,47 +174,12 @@ private:
 
   // spiking history needed by stdp synapses
   std::deque< histentry > history_;
-
-  /*
-   * Structural plasticity
-   */
-
-  // Time of the last update of the Calcium concentration in ms
-  double Ca_t_;
-
-  // Value of the calcium concentration [Ca2+] at Ca_t_. Intracellular calcium
-  // concentration has a linear factor to mean electrical activity of 10^2,
-  // this means, for example, that a [Ca2+] of 0.2 is equivalent to a mean
-  // activity of 20Hz.
-  double Ca_minus_;
-
-  // Time constant for exponential decay of the intracellular calcium
-  // concentration
-  double tau_Ca_;
-
-  // Increase in calcium concentration [Ca2+] for each spike of the neuron
-  double beta_Ca_;
-
-  // Map of the synaptic elements
-  std::map< Name, SynapticElement > synaptic_elements_map_;
 };
 
 inline double
 Archiving_Node::get_spiketime_ms() const
 {
   return last_spike_;
-}
-
-inline double
-Archiving_Node::get_tau_Ca() const
-{
-  return tau_Ca_;
-}
-
-inline double
-Archiving_Node::get_Ca_minus() const
-{
-  return Ca_minus_;
 }
 
 } // of namespace
